calculating a score for a specific tower

+++ Contest/src/contestant/Bucket.java
@@ -3,7 +3,7 @@ package contestant;
 public class Bucket {
     private int id;
     private int angle;
-    private int max_volume;
+    private double max_volume;
 
     public Bucket(int id, int angle, int max_volume){
         this.id = id;
@@ -15,7 +15,7 @@ public class Bucket {
         return angle;
     }
 
-    public int getMax_volume() {
+    public double getMax_volume() {
         return max_volume;
     }
